Reduce AG-UI and knowledge ingest refactor risk
Centralize repeated AG-UI request parsing and injected-tool merge behavior behind narrow shared helpers so future fixes can land in one place without widening pure contract/support module dependencies. Deduplicate knowledge ingest upload classification while preserving the existing prefix retry and skipped-file behavior.

Constraint: Behavior-preserving refactor only; no new dependencies or runtime behavior changes

Rejected: Continue into API module-loader/provider runtime-loader split | broader risk and review surface after safe slices were green

Rejected: Single combined AG-UI helper module | architect review found it pulled tool registry dependencies into pure request/contract modules

Confidence: high

Scope-risk: narrow

Reversibility: clean

Directive: Keep AG-UI request/contract helpers separate from tool/runtime registry imports unless import-boundary tests are updated

Tested: deno fmt --check changed files; DENO_NO_PACKAGE_JSON=1 deno lint changed files; focused AG-UI tests (6 passed, 47 steps); focused knowledge tests (17 passed, 54 steps); deno task typecheck; architect review

Not-tested: Full deno task test suite; API module-loader/provider refactor candidates intentionally deferred
