+++ src/agent/ag-ui-detached-start.ts
@@ -1,13 +1,9 @@
 import { z } from "zod";
 import { INVALID_ARGUMENT } from "#veryfront/errors";
-import { SKILL_TOOL_IDS } from "#veryfront/skill/types.ts";
-import { type Tool, toolRegistry } from "#veryfront/tool";
 import { streamDataStreamEvents } from "./data-stream.ts";
-import {
-  type AgUiInjectedTool,
-  AgUiRequestSchema,
-  normalizeAgUiMessages,
-} from "./ag-ui-host-support.ts";
+import { AgUiRequestSchema, normalizeAgUiMessages } from "./ag-ui-host-support.ts";
+import { extractRequest } from "./ag-ui-request-shared.ts";
+import { type AgUiResumeValue, buildMergedAgUiTools } from "./ag-ui-tool-shared.ts";
 import {
   AgentRuntime,
   RunAlreadyExistsError,
@@ -17,42 +13,12 @@ import type { Agent } from "./types.ts";
 
 const AGENT_ID_PATTERN = /^[a-zA-Z0-9_-]+$/;
 const AG_UI_DETACHED_RUN_ID_SCHEMA = z.string().min(1).max(128).regex(AGENT_ID_PATTERN);
-type AgUiResumeValue = {
-  result: unknown;
-  isError: boolean;
-};
 
 type AgUiContextValue =
   | Record<string, unknown>
   | ((request: Request) => Record<string, unknown> | Promise<Record<string, unknown>>);
 
 type AgUiRuntimePart = Record<string, unknown> & { type: string };
-
-function isRequest(obj: unknown): obj is Request {
-  return (
-    typeof obj === "object" &&
-    obj !== null &&
-    "json" in obj &&
-    typeof obj.json === "function" &&
-    "url" in obj &&
-    typeof obj.url === "string" &&
-    "method" in obj &&
-    typeof obj.method === "string"
-  );
-}
-
-function extractRequest(requestOrCtx: unknown): Request {
-  if (isRequest(requestOrCtx)) return requestOrCtx;
-
-  if (typeof requestOrCtx === "object" && requestOrCtx !== null && "request" in requestOrCtx) {
-    const candidate = (requestOrCtx as Record<string, unknown>).request;
-    if (isRequest(candidate)) return candidate;
-  }
-
-  throw INVALID_ARGUMENT.create({
-    detail: "Invalid handler argument: expected Request or APIContext",
-  });
-}
 
 function buildStreamContext(
   request: AgUiDetachedStartRequest,
@@ -71,66 +37,12 @@ function buildStreamContext(
   };
 }
 
-function createInjectedAgUiTool(
-  runId: string,
-  tool: AgUiInjectedTool,
-  sessionManager: RunResumeSessionManager<AgUiResumeValue>,
-): Tool {
-  return {
-    id: tool.name,
-    type: "function",
-    description: tool.description ?? tool.name,
-    inputSchema: z.record(z.string(), z.unknown()),
-    inputSchemaJson: (tool.parameters ??
-      { type: "object", properties: {}, additionalProperties: true }) as Tool["inputSchemaJson"],
-    execute: async (_input, context) => {
-      const toolCallId = typeof context?.toolCallId === "string" ? context.toolCallId : null;
-      if (!toolCallId) {
-        throw new Error(`Missing toolCallId for injected tool "${tool.name}"`);
-      }
-
-      sessionManager.prepareForSignal(runId, toolCallId);
-      const submitted = await sessionManager.waitForSignal(runId, toolCallId);
-      if (submitted.isError) {
-        throw new Error(
-          typeof submitted.result === "string"
-            ? submitted.result
-            : JSON.stringify(submitted.result),
-        );
-      }
-      return submitted.result;
-    },
-  };
-}
-
 function buildMergedTools(
   agent: Agent,
   request: AgUiDetachedStartRequest,
   sessionManager: RunResumeSessionManager<AgUiResumeValue>,
 ): Agent["config"]["tools"] {
-  const injectedTools = Object.fromEntries(
-    request.tools.map((tool) => [
-      tool.name,
-      createInjectedAgUiTool(request.runId, tool, sessionManager),
-    ]),
-  );
-
-  if (!agent.config.tools) {
-    return Object.keys(injectedTools).length > 0 ? injectedTools : undefined;
-  }
-
-  if (agent.config.tools === true) {
-    const merged: Record<string, Tool | boolean> = {};
-    for (const [toolId] of toolRegistry.getAll()) {
-      if (!agent.config.skills && SKILL_TOOL_IDS.has(toolId)) {
-        continue;
-      }
-      merged[toolId] = true;
-    }
-    return { ...merged, ...injectedTools };
-  }
-
-  return { ...agent.config.tools, ...injectedTools };
+  return buildMergedAgUiTools(agent, request.runId, request.tools, sessionManager);
 }
 
 function scheduleDetachedTask(requestOrCtx: unknown, task: Promise<void>): void {
