+++ src/agent/ag-ui-runtime-handler.ts
@@ -1,7 +1,4 @@
 import { z } from "zod";
-import { INVALID_ARGUMENT } from "#veryfront/errors";
-import { SKILL_TOOL_IDS } from "#veryfront/skill/types.ts";
-import { type Tool, toolRegistry } from "#veryfront/tool";
 import {
   AgentRuntime,
   RunAlreadyExistsError,
@@ -9,10 +6,11 @@ import {
 } from "./runtime/index.ts";
 import type { Agent } from "./types.ts";
 import {
-  type AgUiRuntimeInjectedTool,
   type AgUiRuntimeRequest,
   parseAgUiRuntimeRequestOrError,
 } from "./runtime-ag-ui-contract.ts";
+import { extractRequest } from "./ag-ui-request-shared.ts";
+import { type AgUiResumeValue, buildMergedAgUiTools } from "./ag-ui-tool-shared.ts";
 import { normalizeAgUiRuntimeMessages } from "./ag-ui-runtime-support.ts";
 import {
   createStreamTransformState,
@@ -28,7 +26,6 @@ const AG_UI_HEADERS: Record<string, string> = {
   Connection: "keep-alive",
 };
 
-type AgUiResumeValue = { result: unknown; isError: boolean };
 type AgUiRuntimePart = Record<string, unknown> & { type: string };
 
 export interface AgUiRuntimeLifecycleContext {
@@ -64,34 +61,8 @@ async function invokeLifecycleCallbackAndWait(
   }
 }
 
-function isRequest(obj: unknown): obj is Request {
-  return (
-    typeof obj === "object" &&
-    obj !== null &&
-    "json" in obj &&
-    typeof obj.json === "function" &&
-    "url" in obj &&
-    typeof obj.url === "string" &&
-    "method" in obj &&
-    typeof obj.method === "string"
-  );
-}
-
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-function extractRequest(requestOrCtx: unknown): Request {
-  if (isRequest(requestOrCtx)) return requestOrCtx;
-
-  if (typeof requestOrCtx === "object" && requestOrCtx !== null && "request" in requestOrCtx) {
-    const candidate = (requestOrCtx as Record<string, unknown>).request;
-    if (isRequest(candidate)) return candidate;
-  }
-
-  throw INVALID_ARGUMENT.create({
-    detail: "Invalid handler argument: expected Request or APIContext",
-  });
 }
 
 function buildStreamContext(
@@ -271,66 +242,12 @@ async function createAgUiRuntimeDirectStreamResponse(
   });
 }
 
-function createInjectedAgUiTool(
-  runId: string,
-  tool: AgUiRuntimeInjectedTool,
-  sessionManager: RunResumeSessionManager<AgUiResumeValue>,
-): Tool {
-  return {
-    id: tool.name,
-    type: "function",
-    description: tool.description ?? tool.name,
-    inputSchema: z.record(z.string(), z.unknown()),
-    inputSchemaJson: (tool.parameters ??
-      { type: "object", properties: {}, additionalProperties: true }) as Tool["inputSchemaJson"],
-    execute: async (_input, context) => {
-      const toolCallId = typeof context?.toolCallId === "string" ? context.toolCallId : null;
-      if (!toolCallId) {
-        throw new Error(`Missing toolCallId for injected tool "${tool.name}"`);
-      }
-
-      sessionManager.prepareForSignal(runId, toolCallId);
-      const submitted = await sessionManager.waitForSignal(runId, toolCallId);
-      if (submitted.isError) {
-        throw new Error(
-          typeof submitted.result === "string"
-            ? submitted.result
-            : JSON.stringify(submitted.result),
-        );
-      }
-      return submitted.result;
-    },
-  };
-}
-
 function buildMergedTools(
   agent: Agent,
   request: AgUiRuntimeRequest,
   sessionManager: RunResumeSessionManager<AgUiResumeValue>,
 ): Agent["config"]["tools"] {
-  const injectedTools = Object.fromEntries(
-    request.tools.map((tool) => [
-      tool.name,
-      createInjectedAgUiTool(request.runId, tool, sessionManager),
-    ]),
-  );
-
-  if (!agent.config.tools) {
-    return Object.keys(injectedTools).length > 0 ? injectedTools : undefined;
-  }
-
-  if (agent.config.tools === true) {
-    const merged: Record<string, Tool | boolean> = {};
-    for (const [toolId] of toolRegistry.getAll()) {
-      if (!agent.config.skills && SKILL_TOOL_IDS.has(toolId)) {
-        continue;
-      }
-      merged[toolId] = true;
-    }
-    return { ...merged, ...injectedTools };
-  }
-
-  return { ...agent.config.tools, ...injectedTools };
+  return buildMergedAgUiTools(agent, request.runId, request.tools, sessionManager);
 }
 
 async function createAgUiRuntimeInjectedToolsStreamResponse(
